+++ CHANGELOG.md
@@ -1,3 +1,75 @@
+## [1.1.1](https://github.com/adobe/helix-shared/compare/v1.1.0...v1.1.1) (2019-05-07)
+
+
+### Bug Fixes
+
+* **config:** disallow / in directoryIndex ([2b7a011](https://github.com/adobe/helix-shared/commit/2b7a011)), closes [#106](https://github.com/adobe/helix-shared/issues/106)
+
+# [1.1.0](https://github.com/adobe/helix-shared/compare/v1.0.2...v1.1.0) (2019-05-06)
+
+
+### Bug Fixes
+
+* **proxy:** unify generated names of origins for proxy strains ([5d12ead](https://github.com/adobe/helix-shared/commit/5d12ead))
+
+
+### Features
+
+* **proxy:** enable getting the fastly compatible JSON for an origin ([b1df9d3](https://github.com/adobe/helix-shared/commit/b1df9d3))
+
+## [1.0.2](https://github.com/adobe/helix-shared/compare/v1.0.1...v1.0.2) (2019-05-06)
+
+
+### Bug Fixes
+
+* **package:** update yaml to version 1.5.1 ([de74688](https://github.com/adobe/helix-shared/commit/de74688))
+
+## [1.0.1](https://github.com/adobe/helix-shared/compare/v1.0.0...v1.0.1) (2019-05-03)
+
+
+### Bug Fixes
+
+* include types.js in index.js ([96725f7](https://github.com/adobe/helix-shared/commit/96725f7))
+
+# [1.0.0](https://github.com/adobe/helix-shared/compare/v0.11.0...v1.0.0) (2019-05-03)
+
+
+### Features
+
+* More tools for generic programming ([6c074a3](https://github.com/adobe/helix-shared/commit/6c074a3))
+* Use underscore to mark unused variables ([a760881](https://github.com/adobe/helix-shared/commit/a760881))
+
+
+### BREAKING CHANGES
+
+* Because sequence.js was growing very large,
+that file was split up into multiple components.
+
+The real highlights of this change are:
+- The Trait class
+- The Equals trait
+- The cloning traits
+
+feat(types): Trait – New class to facilitate generic programming. Quite complex. See it's api doc.
+feat(types): Immutable – New trait to mark types as immutable
+feat(types): Equals, eq(), assertEq() – New Trait for equality comparing values in an extensible way
+feat(types): Shallowclone – Trait for shallow cloning values
+feat(types): Deepclone – Trait for recursively cloning values
+feat(types): Pairs, keys(), values() – Trait for iterating over any container as a key/value store
+feat(types): Has, Get, Assign, Delete, Setdefault, Replace – Generic traits for accessing containers
+feat(sequence): seqEq() function to compare sequences
+feat(sequence): The Into trait now Supports typed Arrays, WeakMap and WeakSet
+feat(sequence): The Sequence and Size Trait now supports typed arrays
+feat(dom): Dom Nodes now implement the Equals Trait
+feat(dom): Dom Nodes now implement the Deepclone Trait
+
+# [0.11.0](https://github.com/adobe/helix-shared/compare/v0.10.5...v0.11.0) (2019-05-02)
+
+
+### Features
+
+* **proxystrains:** add support for `path` and `override_host` configuration parameters ([d0c8693](https://github.com/adobe/helix-shared/commit/d0c8693))
+
 ## [0.10.5](https://github.com/adobe/helix-shared/compare/v0.10.4...v0.10.5) (2019-04-29)